Move `_clear_worker_cache` to `MerlinDaskExecutor`
